requirements.txt
Flask==2.3.3
flask-praetorian==1.4.0
py3-validate-email@git+https://github.com/christianrichter/py3-validate-email.git@v1.0.10
waitress==2.1.2
